Sort include directives & forward declarations.

+++ HedgeGI/HedgeGI/BakingFactory.cpp
@@ -1,6 +1,10 @@
 ﻿#include "BakingFactory.h"
-#include "Scene.h"
+
+#include "Bitmap.h"
 #include "Camera.h"
+#include "Light.h"
+#include "Material.h"
+#include "Random.h"
 
 std::mutex BakingFactory::mutex;
 
@@ -27,7 +31,7 @@ Color4 BakingFactory::pathTrace(const RaytracingContext& raytracingContext, cons
     Color3 radiance(0, 0, 0);
     float faceFactor = 1.0f;
 
-    for (int32_t i = 0; i < bakeParams.lightBounceCount; i++)
+    for (int32_t i = 0; i < (int32_t)bakeParams.lightBounceCount; i++)
     {
         const bool shouldApplyBakeParam = !tracingFromEye || i > 0;
 
@@ -36,7 +40,7 @@ Color4 BakingFactory::pathTrace(const RaytracingContext& raytracingContext, cons
 
         // Do russian roulette at highest difficulty fuhuhuhuhuhu
         const float probability = throughput.maxCoeff();
-        if (i > bakeParams.russianRouletteMaxDepth)
+        if (i > (int32_t)bakeParams.russianRouletteMaxDepth)
         {
             if (Random::next() > probability)
                 break;
